Added FastNoise library

+++ src/types.hpp
@@ -28,7 +28,7 @@ namespace ktp {
 
   class KeTerrain;
 
-  using NoiseData      = std::vector<double>;
+  using NoiseData      = std::vector<float>;
   using Pixels         = std::vector<sf::Color>;
   using RawTextureData = std::vector<unsigned char>;
   using TexturePtr     = std::unique_ptr<sf::Texture>;
